added pagination and authentication and authorization

+++ backend/routes/posts.js
@@ -1,6 +1,8 @@
 const express = require('express');
 const multer = require('multer');
 const Post = require('../models/post');
+const checkAuth = require('../middleware/check-auth');
+
 const router = express.Router();
 
 const MIME_TYPE_PATH = {
@@ -25,12 +27,15 @@ const storage = multer.diskStorage({
         cb(null, name +  '-' + Date.now() + '.' + ext);
     }
 })
-router.post('', multer({ storage: storage }).single('image'), (req, res, next) => {
+router.post('',
+    checkAuth, 
+    multer({ storage: storage }).single('image'), (req, res, next) => {
     const url = req.protocol + '://' + req.get('host');
     const post = new Post({
         title: req.body.title,
         content: req.body.content,
-        imagePath: url + '/images/' + req.file.filename
+        imagePath: url + '/images/' + req.file.filename,
+        creator: req.userData.userId
     })
     post.save().then(createdPost => {
         res.status(201).json({
@@ -43,7 +48,7 @@ router.post('', multer({ storage: storage }).single('image'), (req, res, next) =
     });
 });
 
-router.put('/:id', multer({ storage: storage }).single('image'), (req, res, next)=>{
+router.put('/:id', checkAuth, multer({ storage: storage }).single('image'), (req, res, next)=>{
     let imagePath = req.body.imagePath;
     if(req.file){
         const url = req.protocol + '://' + req.get('host');
@@ -53,12 +58,20 @@ router.put('/:id', multer({ storage: storage }).single('image'), (req, res, next
         _id: req.body.id,
         title: req.body.title,
         content: req.body.content,
-        imagePath: imagePath
+        imagePath: imagePath,
+        creator: req.userData.userId
     })
-    Post.updateOne({ _id: req.params.id}, post).then(result => {
-        res.status(200).json({
-            message: 'Update Successful!'
-        });
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+        if(result.nModified > 0){
+            res.status(200).json({
+                message: 'Update Successful!'
+            });
+        } else {
+            res.status(401).json({
+                message: 'Not Authorized!'
+            });
+        }
+        
     });
 });
 
@@ -96,11 +109,17 @@ router.get('/:id', (req, res, next)=>{
     })
 });
 
-router.delete('/:id', (req, res, next)=>{
-    Post.deleteOne({ _id:req.params.id }).then(result => {
-        res.status(200).json({
-            message: 'Post Deleted'
-        });
+router.delete('/:id', checkAuth, (req, res, next)=>{
+    Post.deleteOne({ _id:req.params.id, creator: req.userData.userId }).then(result => {
+        if(result.n > 0){
+            res.status(200).json({
+                message: 'Deletion Successful!'
+            });
+        } else {
+            res.status(401).json({
+                message: 'Not Authorized!'
+            });
+        }
     })
 });
